+++ src/main/java/it/unibz/inf/ontouml/vp/model/ontouml/view/RectangularShape.java
@@ -1,17 +1,17 @@
 package it.unibz.inf.ontouml.vp.model.ontouml.view;
 
 public abstract class RectangularShape extends Shape {
-  Point topLeft = new Point(0.0, 0.0);
-  double width = 20;
-  double height = 10;
+  Point topLeft = new Point(0, 0);
+  int width = 20;
+  int height = 10;
 
-  public RectangularShape(String id, double width, double height) {
+  public RectangularShape(String id, int width, int height) {
     super(id);
     this.width = width;
     this.height = height;
   }
 
-  public RectangularShape(double width, double height) {
+  public RectangularShape(int width, int height) {
     this(null, width, height);
     this.width = width;
     this.height = height;
@@ -25,35 +25,35 @@ public abstract class RectangularShape extends Shape {
     super(null);
   }
 
-  public double getX() {
+  public int getX() {
     return topLeft.getX();
   }
 
-  public void setX(Double x) {
+  public void setX(Integer x) {
     topLeft.setX(x);
   }
 
-  public double getY() {
+  public int getY() {
     return topLeft.getY();
   }
 
-  public void setY(Double y) {
+  public void setY(Integer y) {
     topLeft.setY(y);
   }
 
-  public double getWidth() {
+  public int getWidth() {
     return width;
   }
 
-  public void setWidth(Double width) {
+  public void setWidth(Integer width) {
     this.width = (width != null) ? width : 0;
   }
 
-  public double getHeight() {
+  public int getHeight() {
     return height;
   }
 
-  public void setHeight(Double height) {
+  public void setHeight(Integer height) {
     this.height = (height != null) ? height : 0;
   }
 }
